Move finalized repo to consensus ns
